refactor: domains is now API utils
