+++ pkg/webhook/pod/containerregistry.go
@@ -12,8 +12,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 
+	apiutils "github.com/clastix/capsule/api/utils"
 	capsulev1alpha1 "github.com/clastix/capsule/api/v1alpha1"
-	"github.com/clastix/capsule/api/v1alpha1/domain"
 	capsulewebhook "github.com/clastix/capsule/pkg/webhook"
 	"github.com/clastix/capsule/pkg/webhook/utils"
 )
@@ -49,7 +49,7 @@ func (h *containerRegistryHandler) OnCreate(c client.Client, decoder *admission.
 			var valid, matched bool
 
 			for _, container := range pod.Spec.Containers {
-				registry := domain.NewRegistry(container.Image)
+				registry := apiutils.NewRegistry(container.Image)
 
 				valid = tnt.Spec.ContainerRegistries.ExactMatch(registry.Registry())
